add spot finding pipeline
The aim is to rewrite the different steps of to find spots in a more
modular way, starting with simpler conditions and adding increasing
complexity. It's to have a more general set of tool but also to
personnally understand Peter's code. Many steps use his functions, and some
custom functions are created in a more simple (basic) way than his.